Provide an in-game option to purge data
Allow users of CSIDE to purge game data from within the game. This is slightly experimental in the sense that we're editing the copy of CS used by CSIDE directly, rather than overriding it. I think this has merit moving forwards, as long as the edits are small and well thought out.

Fixes: https://github.com/ChoicescriptIDE/main/issues/49