Resize text via CSS font-size instead of non-standard zoom
CSS zoom is totally non-standard, and Apple keeps applying "interventions" to it as part of their "Desktop-class browsing" initiative.

In iOS 14, with `-webkit-text-size-adjust: none`, iOS re-shrinks italic text back to initial size when the body tries to increase its size with `zoom: 2`. I see no good way to work around this; it's now time to adopt a standards-based approach.

In this commit, a bunch of pixel-based sizes are converting to `ch` and `em` units, and the zoom factor is now a percentage on the body's font-size rather than using CSS zoom.