Don't shrink stat bars for long stat names
Stat bars were using 35ch for their width, but *stat_chart uses smaller font sizes to make long stat names fit, so when the stat name got long, the font would be smaller and the bar would shrink, causing wrapping.

Now we declare the stat bar size in terms of rems, so it won't shrink when any particular bar's font size changes.

(To do this, we had to move the zoom factor to the HTML documentElement element, so changes in zoom factor would adjust rem.)