Use only one object store
There's no way to create an object store except when creating/upgrading.

If we use a separate object store per ChoiceScript `storeName`, there's no way to force an upgrade event to create the store.

So we'll just use one object store, and ensure unique keys, like we do in localStorage.